A 1st patch from Pierre Smits for <<Have the ability to test solr queries through widgets (screens & forms)>> https://issues.apache.org/jira/browse/OFBIZ-6595
Adding widgest (screens, forms, etc) so that an end-user can execute a search  against the data indexed would enhance the applicability of the application.
This patch is the first increment to address the issue. Please feel free to test drive and provide feedback
Currently it includes the following aspects:

    SolrMenus.xml - providing the application menu, items:
        main - opening the start screen of the app
        admin - opening the Solr admin page in a new target
        newquery - opening the query screen
    SolrScreens.xml - defining Solr specific screen widgets
    CommonScreens.xml - defining the common screens for Solr specific screen widgets
    SolrUiLabels.xml - defining Solr app specific labels
    SolrForms.xml - defining Solr specific form widgets
    controller.xml - (re)defining request-map and view-map re widgets

This file also includes a modification to web.xml of the Solr component, changing the mainDecoratorLocation
This patch currently doesn't include functionality to execute a query through widgets and show the result.


git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk/specialpurpose/solr@1736878 13f79535-47bb-0310-9956-ffa450edef68