Implemented: removed the "controller" related logic, that is now provided by the new ControlFilter, from the ContextFilter; modified existing applications to use the two filters in a chain; removed from the other specialized filters all the logic that was duplicated or extended from ContextFilter.
A web application, in order to leverage the OFBiz framework, requires that a 
series of objects are in its contexts (servlet context, session and request) 
such as "delegator", "delegatorName", "dispatcher", "security" etc. etc...
This setup is performed by the logic contained in the servlet filter implemented 
by the ContextFilter class.
The execution of this logic is required for the application to run properly.
However, before this commit, in the ContextFilter there was other logic, related 
to access control and redirection rules (some of them performed in coordination 
with the ControlServlet), making it difficult to deploy this filter in all the 
web applications, especially the ones that implement special handling of paths.
In fact, this filter was deployed in most but not all the web application in the 
OFBiz codebase: specifically it was not deployed in web applications that 
require the execution of other filters (e.g. CatalogUrlFilter, etc...) like the 
ones in the "ecommerce" and "solr" components.



git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk/specialpurpose/solr@1761304 13f79535-47bb-0310-9956-ffa450edef68